+++ test/mesh/provisioning_provisioner_test.cpp
@@ -41,7 +41,7 @@
 #include <string.h>
 #include "mesh/pb_adv.h"
 #include "mesh/pb_gatt.h"
-#include "ble/gatt-service/mesh_provisioning_service_server.h"
+#include "mesh/gatt-service/mesh_provisioning_service_server.h"
 #include "mesh/provisioning.h"
 #include "mesh/provisioning_provisioner.h"
 #include "hci_dump.h"